Agent foundation setup
Agent initialise a flask app to run on , then it waits for the load balancer to send a request on a port then sends a public key to the load balancer for a key based security system.  Handlers.py handles the linking to the load balancer. Health check sends a latency check to the load balancer which is then used in resource monitor to see what status the server is in. Resource monitor is not completed but has basic functionality.