Backend data into 2 charts and alert table data into a frontend table
Created a response factory to generate a response but to make this fast as it was taking roughly 10s to get to frontend it created a alert manager file to manage server alerts and maintain a cached list of alerts. Alert manager uses threading to update alerts in the background making it seconds faster. Removed critical from the health check